use pi to verify user payment instead of dcrdata
